requirements.txt
# open3d
tqdm
six
numpy
point-cloud-utils==0.27.0
wandb
matplotlib==3.6.2
seaborn==0.13.0
pandas==1.5.3
pyvista
omegaconf==2.2.3
scikit-learn
torchmetrics
reformer-pytorch
chardet
networkx
imageio[ffmpeg]